Organize tests and compiler issues

+++ src/eskilib_trie.c
@@ -17,7 +17,7 @@ struct eskilib_Trie* eskilib_trie_malloc() {
 void eskilib_trie_free(struct eskilib_Trie* tree) {
 	assert(tree != NULL);
 
-	for (uint_fast32_t i = 0; i < ESKILIB_TRIE_LETTERS; i++) {
+	for (size_t i = 0; i < ESKILIB_TRIE_LETTERS; i++) {
 		if (tree->nodes[i] != NULL)
 			eskilib_trie_free(tree->nodes[i]);
 		else
@@ -26,14 +26,14 @@ void eskilib_trie_free(struct eskilib_Trie* tree) {
 	free(tree);
 }
 
-void eskilib_trie_add(char* string, uint_fast32_t length, struct eskilib_Trie* tree) {
+void eskilib_trie_add(char* string, size_t length, struct eskilib_Trie* tree) {
 	assert(string != NULL);
 	assert(length > 0);
 	assert(tree != NULL);
 
 	int index = 0;
 
-	for (uint_fast32_t i = 0; i < length - 1; i++) { //string.length - 1 because it includes null terminator
+	for (size_t i = 0; i < length - 1; i++) { //string.length - 1 because it includes null terminator
 		index = (int)string[i] - ' ';
 		if (tree->nodes[index] == NULL) {
 			tree->nodes[index] = calloc(1, sizeof(struct eskilib_Trie));
@@ -54,7 +54,7 @@ void eskilib_trie_add_string(struct eskilib_String string, struct eskilib_Trie* 
 
 	int index = 0;
 
-	for (uint_fast32_t i = 0; i < string.length - 1; i++) { //string.length - 1 because it includes null terminator
+	for (size_t i = 0; i < string.length - 1; i++) { //string.length - 1 because it includes null terminator
 		index = (int)string.value[i] - ' ';
 		if (tree->nodes[index] == NULL) {
 			tree->nodes[index] = calloc(1, sizeof(struct eskilib_Trie));
@@ -68,20 +68,20 @@ void eskilib_trie_add_string(struct eskilib_String string, struct eskilib_Trie* 
 	tree->is_end_of_a_word = true;
 }
 
-void eskilib_trie_add_multiple(struct eskilib_String* strings, uint_fast32_t count, struct eskilib_Trie* tree) {
-	for (uint_fast32_t i = 0; i < count; i++) {
+void eskilib_trie_add_multiple(struct eskilib_String* strings, int count, struct eskilib_Trie* tree) {
+	for (int i = 0; i < count; i++) {
 		eskilib_trie_add_string(strings[i], tree);
 	}
 }
 
-struct eskilib_Trie* eskilib_trie_search(char* string, uint_fast32_t length, struct eskilib_Trie* tree) {
+struct eskilib_Trie* eskilib_trie_search(char* string, size_t length, struct eskilib_Trie* tree) {
 	assert(string != NULL);
 	assert(length > 0);
 	assert(tree != NULL);
 
 	int index = 0;
 
-	for (uint_fast32_t i = 0; i < length - 1; i++) {
+	for (size_t i = 0; i < length - 1; i++) {
 		index = (int)string[i] - ' ';
 		if (tree->nodes[index] == NULL)
 			return NULL;
@@ -102,7 +102,7 @@ struct eskilib_Trie* eskilib_trie_search_string(struct eskilib_String string, st
 
 	int index = 0;
 
-	for (uint_fast32_t i = 0; i < string.length - 1; i++) {
+	for (size_t i = 0; i < string.length - 1; i++) {
 		index = (int)string.value[i] - ' ';
 		if (tree->nodes[index] == NULL)
 			return NULL;
@@ -117,8 +117,8 @@ struct eskilib_Trie* eskilib_trie_search_string(struct eskilib_String string, st
 }
 
 void eskilib_trie_match(char* matches[],
-				uint_fast32_t* string_position,
-				uint_fast32_t* matches_position,
+				size_t* string_position,
+				size_t* matches_position,
 				const uint_fast32_t max_match_length,
 				struct eskilib_Trie* tree) {
 	for (uint_fast32_t i = 0; i < ESKILIB_TRIE_LETTERS; i++) {
@@ -149,17 +149,17 @@ void eskilib_trie_match(char* matches[],
 	}
 }
 
-uint_fast32_t eskilib_trie_matches(char* matches[], const uint_fast32_t max_match_length, struct eskilib_Trie* tree) {
-	uint_fast32_t string_position = 0;
-	uint_fast32_t matches_position = 0;
+size_t eskilib_trie_matches(char* matches[], const uint_fast32_t max_match_length, struct eskilib_Trie* tree) {
+	size_t string_position = 0;
+	size_t matches_position = 0;
 
 	eskilib_trie_match(matches, &string_position, &matches_position, max_match_length, tree);
 
 	return matches_position;
 }
 
-uint_fast32_t eskilib_trie_get(char* search,
-				       uint_fast32_t search_length,
+size_t eskilib_trie_get(char* search,
+				       size_t search_length,
 				       char* matches[],
 				       const uint_fast32_t max_match_length,
 				       struct eskilib_Trie* tree) {
